fix: 채팅 세션 중계 기능 gateway로 이전
