+++ server/src/api/chat/chat.gateway.spec.ts
@@ -1,13 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ChatGateway } from './chat.gateway';
-import { ChatService } from './chat.service';
 
 describe('ChatGateway', () => {
   let gateway: ChatGateway;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [ChatGateway, ChatService],
+      providers: [ChatGateway],
     }).compile();
 
     gateway = module.get<ChatGateway>(ChatGateway);
